Remove `clone_repo`, `repo_exists` function (integrate into RepoProfile); Remove `swesmith.utils`
