+++ tests/harness/test_harness_utils.py
@@ -1,7 +1,36 @@
 from swesmith.harness.utils import *
+from swesmith.profiles import RepoProfile
+import os
 
 
-def test_get_test_paths(tmp_path):
+class MockRepoProfile(RepoProfile):
+    """Mock RepoProfile for testing that uses a local directory instead of cloning."""
+
+    def __init__(self, test_dir: str):
+        self.owner = "test"
+        self.repo = "test_repo"
+        self.commit = "test12345678"
+        self._test_dir = test_dir
+
+    def build_image(self):
+        pass
+
+    def log_parser(self, log: str) -> dict[str, str]:
+        return {}
+
+    def clone(self, dest: str | None = None) -> str | None:
+        """Override clone to use the test directory instead of git clone."""
+        dest = self.repo_name if not dest else dest
+        if not os.path.exists(dest):
+            # Copy the test directory to the expected repo name
+            import shutil
+
+            shutil.copytree(self._test_dir, dest)
+            return dest
+        return None
+
+
+def test_get_cached_test_paths(tmp_path):
     # Create directory structure
     (tmp_path / "tests").mkdir()
     (tmp_path / "src").mkdir()
@@ -24,8 +53,11 @@ def test_get_test_paths(tmp_path):
         f.parent.mkdir(parents=True, exist_ok=True)
         f.write_text("# test file" if f in test_files else "# not a test file")
 
-    # Call get_test_paths
-    result = get_test_paths(str(tmp_path))
+    # Create mock RepoProfile
+    mock_rp = MockRepoProfile(str(tmp_path))
+
+    # Call get_cached_test_paths
+    result = get_cached_test_paths(mock_rp)
     result_set = set(str(p) for p in result)
     # Expected: all test_files, relative to tmp_path
     expected = set(str(f.relative_to(tmp_path)) for f in test_files)
