+++ tests/test_profiles.py
@@ -1,6 +1,8 @@
 from swesmith.constants import ORG_NAME_GH
 from swesmith.profiles import global_registry, RepoProfile
 from swesmith.profiles.utils import INSTALL_CMAKE, INSTALL_BAZEL
+from unittest.mock import patch
+import subprocess
 
 
 def test_registry_keys_and_lookup():
@@ -26,6 +28,61 @@ def test_image_name():
     assert repo_profile.owner in image_name
     assert repo_profile.repo in image_name
     assert repo_profile.commit[:8] in image_name
+
+
+def test_repo_profile_clone():
+    """Test the RepoProfile.clone method, adapted from the original clone_repo test."""
+    repo_profile = global_registry.get("mewwts__addict.75284f95")
+
+    # Test with default dest (should use repo_name)
+    expected_dest = repo_profile.repo_name
+    expected_cmd = f"git clone git@github.com:{repo_profile.mirror_name}.git {repo_profile.repo_name}"
+
+    with (
+        patch("os.path.exists", return_value=False) as mock_exists,
+        patch("subprocess.run") as mock_run,
+    ):
+        result = repo_profile.clone()
+        mock_exists.assert_called_once_with(expected_dest)
+        mock_run.assert_called_once_with(
+            expected_cmd,
+            check=True,
+            shell=True,
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+        )
+        assert result == expected_dest
+
+    # Test with custom dest specified
+    custom_dest = "some_dir"
+    expected_cmd_with_dest = (
+        f"git clone git@github.com:{repo_profile.mirror_name}.git {custom_dest}"
+    )
+
+    with (
+        patch("os.path.exists", return_value=False) as mock_exists,
+        patch("subprocess.run") as mock_run,
+    ):
+        result = repo_profile.clone(custom_dest)
+        mock_exists.assert_called_once_with(custom_dest)
+        mock_run.assert_called_once_with(
+            expected_cmd_with_dest,
+            check=True,
+            shell=True,
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+        )
+        assert result == custom_dest
+
+    # Test when repo already exists
+    with (
+        patch("os.path.exists", return_value=True) as mock_exists,
+        patch("subprocess.run") as mock_run,
+    ):
+        result = repo_profile.clone(custom_dest)
+        mock_exists.assert_called_once_with(custom_dest)
+        mock_run.assert_not_called()
+        assert result is None
 
 
 def test_python_log_parser():
